+++ MonkeyTV/Controllers/PlayerViewController.swift
@@ -9,7 +9,7 @@ import UIKit
 import youtube_ios_player_helper
 
 class PlayerViewController: UIViewController {
-    
+    var videoId: String = "FjJtmJteK58"
     private var landscapeConstraints: [NSLayoutConstraint] = []
     private var portraitConstraints: [NSLayoutConstraint] = []
     // MARK: - support
@@ -17,52 +17,57 @@ class PlayerViewController: UIViewController {
     private let smallSymbolConfig = UIImage.SymbolConfiguration(pointSize: 20)
     private var initialY: CGFloat = 0
     private var finalY: CGFloat = 0
+    private var videoDuration = 0
+    private var timer: Timer?
+    private var bulletChats = [BulletChat]()
     // MARK: - Bools
     private var isPanning = false
     private var isDanMuDisplayed = false
     private var danMuTextFiedIsShow = false
     private var videoIsPlaying = true
     private var playerIsShrink = false
+    private var isSideViewVisible = false
     // MARK: - Views
-    private var ytVideoPlayerView = {
+    private var ytVideoPlayerView: YTPlayerView = {
         let view = YTPlayerView()
         view.translatesAutoresizingMaskIntoConstraints = false
         return view
     }()
-    private var buttonsView = {
+    private var danmuView: DanMuView = {
+        let view = DanMuView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+    private var buttonsView: UIView = {
         let view = UIView()
         view.translatesAutoresizingMaskIntoConstraints = false
         return view
     }()
-    private var tableView = {
+    private var tableView: UITableView = {
         var tableView = UITableView()
         tableView.rowHeight = UITableView.automaticDimension
         tableView.separatorStyle = .none
         tableView.allowsSelection = false
-        tableView.backgroundColor = .systemMint
         tableView.register(ChatroomButtonTableViewCell.self,
                            forCellReuseIdentifier:
                             ChatroomButtonTableViewCell.identifier)
         tableView.translatesAutoresizingMaskIntoConstraints = false
         return tableView
     }()
+    private var snapshot = NSDiffableDataSourceSnapshot<OneSection, MKShow>()
+    private var dataSource: UITableViewDiffableDataSource<OneSection, MKShow>!
     // MARK: - Buttons
     private lazy var changeOrientationButton = {
         return UIButton.createPlayerButton(
             image: UIImage.systemAsset(.enlarge, configuration: smallSymbolConfig),
-            color: .white, cornerRadius: 15)
+            color: .white, cornerRadius: 10)
     }()
     private lazy var showDanMuButton = {
         let button = UIButton.createPlayerButton(
             image: UIImage.systemAsset(.square, configuration: smallSymbolConfig),
-            color: .white, cornerRadius: 12)
+            color: .white, cornerRadius: 10)
         button.setTitle("彈幕", for: .normal)
         return button
-    }()
-    private lazy var showDanMuTextFieldButton = {
-        return UIButton.createPlayerButton(
-            image: UIImage.systemAsset(.submitDanMu, configuration: smallSymbolConfig),
-            color: .white, cornerRadius: 15)
     }()
     private lazy var pauseButton = {
         return UIButton.createPlayerButton(
@@ -79,33 +84,11 @@ class PlayerViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        let playerVars: [AnyHashable: Any] = [
-            "margin": 0,
-            "width": 100,
-            //            "height": 100,
-            "frameborder": 0,
-            "loop": 0,
-            "playsigline": 1,
-            "controls": 0,
-            //                "autohide": 1,
-            "showinfo": 0,
-            "fs": 0,
-            //            "rel": 0,
-            "autoplay": 1
-        ]
-        ytVideoPlayerView.load(withVideoId: "FjJtmJteK58", playerVars: playerVars)
-        
-        
-        setup()
+        setupUILayout()
+        setupTableView()
+        getDanMuData()
+        setupVideoLauncher()
     }
-    
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-        if UIDevice.current.orientation.isLandscape {
-            
-        }
-    }
-    
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
         super.viewWillTransition(to: size, with: coordinator)
         
@@ -119,11 +102,12 @@ class PlayerViewController: UIViewController {
                 NSLayoutConstraint.deactivate(landscapeConstraints)
                 NSLayoutConstraint.activate(portraitConstraints)
             }
-        }, completion: { context in
-            
+        }, completion: { _ in
         })
     }
-    
+    deinit {
+        timer = nil
+    }
     // MARK: - Button View Gesture
     private func addButtonViewGesture() {
         let singleFinger = UITapGestureRecognizer(target: self, action: #selector(showButtonView))
@@ -136,13 +120,11 @@ class PlayerViewController: UIViewController {
     @objc func showButtonView() {
         buttonsView.backgroundColor = UIColor(white: 0, alpha: 0.3)
         showDanMuButton.isHidden = false
-        showDanMuTextFieldButton.isHidden = false
         pauseButton.isHidden = false
         videoSlider.isHidden = false
         changeOrientationButton.isHidden = false
         DispatchQueue.main.asyncAfter(deadline: .now() + 5.0) { [weak self] in
             self?.showDanMuButton.isHidden = true
-            self?.showDanMuTextFieldButton.isHidden = true
             self?.pauseButton.isHidden = true
             self?.videoSlider.isHidden = true
             self?.changeOrientationButton.isHidden = true
@@ -150,78 +132,122 @@ class PlayerViewController: UIViewController {
         }
     }
     @objc func handlePanGesture(_ gesture: UIPanGestureRecognizer) {
-           let translation = gesture.translation(in: buttonsView)
-           
-           switch gesture.state {
-           case .began:
-               initialY = self.view.frame.origin.y
-               finalY = initialY + translation.y
-               isPanning = true
-           case .changed:
-               finalY = initialY + translation.y
-               // 判斷是否超過 1/3 並且啟始點 Y 座標小於最終 Y 座標
-               if finalY > initialY && (finalY - initialY) > self.view.frame.height / 3 {
-                   // 在此處執行縮小視圖控制器的操作
-                   // 例如，改變視圖控制器的 frame
-                   self.view.frame.origin.y = finalY
-               }
-           case .ended, .cancelled:
-               isPanning = false
-               // 判斷是否具有一定加速度並且手勢已經達到 1/3 以上
-               let velocity = gesture.velocity(in: self.view)
-               if finalY > initialY && (finalY - initialY) > self.view.frame.height / 3 && velocity.y > 100 {
-                   // 在此處執行往下縮小視圖控制器的操作
-                   // 例如，使用動畫將視圖控制器縮小到底部
-                   dismiss(animated: true)
-                   UIView.animate(withDuration: 0.3) {
-                       self.view.frame.origin.y = self.view.frame.height
-                   }
-               } else {
-                   // 恢復視圖控制器到原始位置
-                   UIView.animate(withDuration: 0.3) {
-                       self.view.frame.origin.y = 0
-                   }
-               }
-           default:
-               break
-           }
-       }
+        let translation = gesture.translation(in: buttonsView)
+        
+        switch gesture.state {
+        case .began:
+            initialY = self.view.frame.origin.y
+            finalY = initialY + translation.y
+            isPanning = true
+        case .changed:
+            finalY = initialY + translation.y
+            // 判斷是否超過 1/3 並且啟始點 Y 座標小於最終 Y 座標
+            if finalY > initialY && (finalY - initialY) > self.view.frame.height / 3 {
+                // 在此處執行縮小視圖控制器的操作
+                // 例如，改變視圖控制器的 frame
+                self.view.frame.origin.y = finalY
+            }
+        case .ended, .cancelled:
+            isPanning = false
+            // 判斷是否具有一定加速度並且手勢已經達到 1/3 以上
+            let velocity = gesture.velocity(in: self.view)
+            if finalY > initialY && (finalY - initialY) > self.view.frame.height / 3 && velocity.y > 100 {
+                // 在此處執行往下縮小視圖控制器的操作
+                // 例如，使用動畫將視圖控制器縮小到底部
+                dismiss(animated: true)
+                UIView.animate(withDuration: 0.3) {
+                    self.view.frame.origin.y = self.view.frame.height
+                }
+            } else {
+                // 恢復視圖控制器到原始位置
+                UIView.animate(withDuration: 0.3) {
+                    self.view.frame.origin.y = 0
+                }
+            }
+        default:
+            break
+        }
+    }
     // MARK: - Buttons AddTaget
     private func setBtnsAddtarget() {
-//        showDanMuButton.addTarget(self, action: #selector(showDanMuView(sender:)),
-//                                  for: .touchUpInside)
-//        showDanMuTextFieldButton.addTarget(self, action: #selector(showDanMuTextField(sender:)),
-//                                           for: .touchUpInside)
-//        pauseButton.addTarget(self, action: #selector(pauseVideo(sender:)), for: .touchUpInside)
+        showDanMuButton.addTarget(self, action: #selector(showDanMuView(sender:)),
+                                  for: .touchUpInside)
+        pauseButton.addTarget(self, action: #selector(pauseVideo(sender:)), for: .touchUpInside)
         changeOrientationButton.addTarget(self,
                                           action: #selector(changeOrientation(sender:)),
                                           for: .touchUpInside)
-//        submitDanMuButton.addTarget(self, action: #selector(submitMyDanMuButton(sender:)),
-//                                    for: .touchUpInside)
-//        videoSlider.addTarget(self, action: #selector(handleSliderChange(sender:)),
-//                              for: .valueChanged)
+        videoSlider.addTarget(self, action: #selector(handleSliderChange(sender:)),
+                              for: .valueChanged)
+    }
+    @objc func showDanMuRoom(sender: UISlider) {
+        if isSideViewVisible {
+            // 關閉視圖
+            UIView.animate(withDuration: 0.3) {
+                self.danmuView.frame.origin.x = -200
+            }
+        } else {
+            // 彈出視圖
+            UIView.animate(withDuration: 0.3) {
+                self.danmuView.frame.origin.x = 0
+            }
+        }
+        isSideViewVisible = !isSideViewVisible
+    }
+    
+    // MARK: - Handle Slider Change
+    @objc func handleSliderChange(sender: UISlider) {
+        let desiredTime = sender.value
+        ytVideoPlayerView.seek(toSeconds: desiredTime, allowSeekAhead: true)
+    }
+    // MARK: - pauseVideo
+    @objc func pauseVideo(sender: UIButton) {
+        danmuView.isPause = !danmuView.isPause
+        if videoIsPlaying {
+            sender.setImage(UIImage.systemAsset(.play, configuration: self.symbolConfig),
+                            for: .normal)
+            ytVideoPlayerView.pauseVideo()
+        } else {
+            sender.setImage(UIImage.systemAsset(.pause, configuration: self.symbolConfig),
+                            for: .normal)
+            ytVideoPlayerView.playVideo()
+        }
+        videoIsPlaying.toggle()
     }
     // MARK: - Change Orientation
     @objc func changeOrientation(sender: UIButton) {
         if playerIsShrink == false {
             tableView.removeFromSuperview()
+            sender.setImage(UIImage.systemAsset(.shrink, configuration: self.symbolConfig), for: .normal)
             NSLayoutConstraint.deactivate(portraitConstraints)
             NSLayoutConstraint.activate(landscapeConstraints)
         } else {
             view.addSubview(tableView)
+            sender.setImage(UIImage.systemAsset(.enlarge, configuration: self.symbolConfig), for: .normal)
             NSLayoutConstraint.deactivate(landscapeConstraints)
             NSLayoutConstraint.activate(portraitConstraints)
         }
         playerIsShrink.toggle()
     }
+    // MARK: - showDanMuView
+    @objc func showDanMuView(sender: UIButton) {
+        if !isDanMuDisplayed {
+            sender.setImage(UIImage.systemAsset(.checkmarkSquare, configuration: smallSymbolConfig), for: .normal)
+            danmuView.isHidden = false
+        } else {
+            sender.setImage(UIImage.systemAsset(.square, configuration: smallSymbolConfig), for: .normal)
+            danmuView.isHidden = true
+        }
+        isDanMuDisplayed.toggle()
+    }
     // MARK: -
-    func setup() {
+    private func setupUILayout() {
         setBtnsAddtarget()
         view.addSubview(ytVideoPlayerView)
         view.addSubview(tableView)
+        view.addSubview(danmuView)
         view.addSubview(buttonsView)
+        
         buttonsView.addSubview(showDanMuButton)
-        //        buttonsView.addSubview(showDanMuTextFieldButton)
         buttonsView.addSubview(pauseButton)
         buttonsView.addSubview(videoSlider)
         buttonsView.addSubview(changeOrientationButton)
@@ -229,36 +255,44 @@ class PlayerViewController: UIViewController {
         addButtonViewGesture()
         
         landscapeConstraints = [
-            ytVideoPlayerView.leadingAnchor.constraint(equalTo: self.view.leadingAnchor),
+            ytVideoPlayerView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             ytVideoPlayerView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
             ytVideoPlayerView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            ytVideoPlayerView.bottomAnchor.constraint(
-                equalTo: view.bottomAnchor),
+            ytVideoPlayerView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
             
             buttonsView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             buttonsView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
             buttonsView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            buttonsView.heightAnchor.constraint(equalTo: view.widthAnchor, multiplier: 9 / 16),
+            buttonsView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            
+            danmuView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            danmuView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            danmuView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            danmuView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            //
+            //            danMuRoomView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            //            danMuRoomView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            //            danMuRoomView.heightAnchor.constraint(equalTo: view.heightAnchor, multiplier: 1/6),
+            //            danMuRoomView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
             
             pauseButton.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor),
             pauseButton.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor),
             pauseButton.heightAnchor.constraint(equalToConstant: 60),
             pauseButton.widthAnchor.constraint(equalToConstant: 60),
             
-            changeOrientationButton.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor,constant: -50),
-            changeOrientationButton.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor),
-            changeOrientationButton.heightAnchor.constraint(equalToConstant: 60),
-            changeOrientationButton.widthAnchor.constraint(equalToConstant: 60),
-
-            showDanMuButton.centerXAnchor.constraint(equalTo: buttonsView.centerXAnchor, constant: 180),
-            showDanMuButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -120),
+            showDanMuButton.centerXAnchor.constraint(equalTo: buttonsView.centerXAnchor, constant: 80),
+            showDanMuButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -32),
             showDanMuButton.widthAnchor.constraint(equalToConstant: 90),
             showDanMuButton.heightAnchor.constraint(equalToConstant: 30),
             
+            changeOrientationButton.leadingAnchor.constraint(equalTo: showDanMuButton.trailingAnchor, constant: 12),
+            changeOrientationButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -32),
+            changeOrientationButton.heightAnchor.constraint(equalToConstant: 30),
+            changeOrientationButton.widthAnchor.constraint(equalToConstant: 30),
             
             videoSlider.leadingAnchor.constraint(equalTo: buttonsView.leadingAnchor),
             videoSlider.trailingAnchor.constraint(equalTo: buttonsView.trailingAnchor),
-            videoSlider.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -160),
+            videoSlider.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -80),
             videoSlider.heightAnchor.constraint(equalToConstant: 10)
         ]
         
@@ -278,22 +312,24 @@ class PlayerViewController: UIViewController {
             buttonsView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             buttonsView.heightAnchor.constraint(equalTo: view.widthAnchor, multiplier: 9 / 16),
             
+            danmuView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            danmuView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            danmuView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            danmuView.heightAnchor.constraint(equalTo: view.widthAnchor, multiplier: 9 / 16),
+            
             pauseButton.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor),
             pauseButton.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor),
             pauseButton.heightAnchor.constraint(equalToConstant: 60),
             pauseButton.widthAnchor.constraint(equalToConstant: 60),
             
-            showDanMuButton.centerXAnchor.constraint(equalTo: buttonsView.centerXAnchor, constant: 40),
+            showDanMuButton.centerXAnchor.constraint(equalTo: buttonsView.centerXAnchor),
             showDanMuButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -8),
             showDanMuButton.widthAnchor.constraint(equalToConstant: 90),
             showDanMuButton.heightAnchor.constraint(equalToConstant: 30),
-            
-            
-            changeOrientationButton.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor,constant: -50),
-            changeOrientationButton.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor),
-            changeOrientationButton.heightAnchor.constraint(equalToConstant: 60),
-            changeOrientationButton.widthAnchor.constraint(equalToConstant: 60),
-
+            changeOrientationButton.leadingAnchor.constraint(equalTo: showDanMuButton.trailingAnchor, constant: 12),
+            changeOrientationButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -8),
+            changeOrientationButton.heightAnchor.constraint(equalToConstant: 30),
+            changeOrientationButton.widthAnchor.constraint(equalToConstant: 30),
             
             videoSlider.leadingAnchor.constraint(equalTo: buttonsView.leadingAnchor),
             videoSlider.trailingAnchor.constraint(equalTo: buttonsView.trailingAnchor),
@@ -301,81 +337,157 @@ class PlayerViewController: UIViewController {
             videoSlider.heightAnchor.constraint(equalToConstant: 10)
         ]
         NSLayoutConstraint.activate(portraitConstraints)
-        
+    }
+    private func setupVideoLauncher() {
+        buttonsView.backgroundColor = UIColor(white: 0, alpha: 0.0)
+        //        danMuTextField.isHidden = true
+        //        submitDanMuButton.isHidden = true
+        changeOrientationButton.isHidden = true
+        showDanMuButton.isHidden = true
+        pauseButton.isHidden = true
+        videoSlider.isHidden = true
+        setBtnsAddtarget()
+        setDanMu()
+        ytVideoPlayerView.delegate = self
+        ytVideoPlayerView.backgroundColor = .black
+        addButtonViewGesture()
+        let playerVars: [AnyHashable: Any] = [
+            "margin": 0,
+            "width": 100,
+            //            "height": 100,
+            "frameborder": 0,
+            "loop": 0,
+            "playsigline": 1,
+            "controls": 0,
+            //                "autohide": 1,
+            "showinfo": 0,
+            "fs": 0,
+            //            "rel": 0,
+            "autoplay": 1
+        ]
+        ytVideoPlayerView.load(withVideoId: videoId, playerVars: playerVars)
+    }
+    // MARK: - Dan Mu
+    private func setDanMu() {
+        danmuView.isHidden = true
+        danmuView.minSpeed = 1
+        danmuView.maxSpeed = 2
+        danmuView.gap = 20
+        danmuView.lineHeight = 30
+        danmuView.start()
+        timer = Timer.scheduledTimer(timeInterval: 0.4,
+                                     target: self, selector: #selector(addDanMuText),
+                                     userInfo: nil, repeats: false)
+    }
+    @objc func addDanMuText() {
+        let interval = CGFloat.random(in: 0.3...1.0)
+        Timer.scheduledTimer(timeInterval: interval,
+                             target: self, selector: #selector(addDanMuText),
+                             userInfo: nil, repeats: false)
+    }
+}
+// MARK: - YTPlayerViewDelegate
+extension PlayerViewController: YTPlayerViewDelegate {
+    func playerViewDidBecomeReady(_ playerView: YTPlayerView) {
+        getVideoDuratiion()
+    }
+    func playerView(_ playerView: YTPlayerView, didChangeTo state: YTPlayerState) {
+        switch state {
+        case .playing:
+            print("Video is playing")
+        case .paused:
+            print("Video is paused")
+        case .ended:
+            print("Video playback ended")
+        default:
+            break
+        }
+    }
+    func playerView(_ playerView: YTPlayerView, receivedError error: YTPlayerError) {
+    }
+    //    func playerViewPreferredInitialLoading(_ playerView: YTPlayerView) -> UIView? {
+    //    }
+    func playerView(_ playerView: YTPlayerView, didPlayTime playTime: Float) {
+        self.bulletChats.sort()
+        bulletChats.forEach({
+            if playTime >= $0.popTime {
+                danmuView.danmuQueue.append(($0.content, false))
+                self.bulletChats.remove(at: 0)
+            }
+        })
+        if self.videoDuration != 0 {
+            self.videoSlider.maximumValue = Float(self.videoDuration)
+            videoSlider.value = playTime
+        }
+    }
+    func getVideoDuratiion() {
+        ytVideoPlayerView.duration { (duration, error) in
+            if let error = error {
+                print("無法取得影片總時間：\(error.localizedDescription)")
+            } else {
+                self.videoDuration = Int(duration)
+            }
+        }
+    }
+    // MARK: - getDanMuData
+    private func getDanMuData() {
+        FirestoreManager.bulletChat.whereField(
+            "videoId", isEqualTo: videoId).getDocuments { querySnapshot, error in
+                if let querySnapshot = querySnapshot {
+                    for document in querySnapshot.documents {
+                        do {
+                            let jsonData = try JSONSerialization.data(withJSONObject: document.data())
+                            let decodedObject = try JSONDecoder().decode(BulletChatData.self, from: jsonData)
+                            self.bulletChats.append(decodedObject.bulletChat)
+                        } catch {
+                            print("\(error)")
+                        }
+                    }
+                }
+            }
     }
 }
 
-
-
-
-
-//    private func setLandscapeAutoLayOut() {
-//        tableView.removeFromSuperview()
-//        removeAllContraints()
-//        NSLayoutConstraint.activate([
-//
-//            danmuView.centerXAnchor.constraint(equalTo: rootViewController.view.centerXAnchor),
-//            danmuView.centerYAnchor.constraint(equalTo: rootViewController.view.centerYAnchor),
-//            danmuView.widthAnchor.constraint(equalTo: rootViewController.view.heightAnchor),
-//            danmuView.heightAnchor.constraint(equalTo: rootViewController.view.widthAnchor),
-//
-//            videoSlider.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor),
-//            videoSlider.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor, constant: 100),
-//            videoSlider.widthAnchor.constraint(equalTo: rootViewController.view.heightAnchor, constant: -450),
-//            videoSlider.heightAnchor.constraint(equalToConstant: 50),
-//            changeOrientationButton.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor),
-//            changeOrientationButton.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor, constant: 150),
-//            changeOrientationButton.widthAnchor.constraint(equalToConstant: 40),
-//            changeOrientationButton.heightAnchor.constraint(equalToConstant: 40),
-//            showDanMuButton.centerXAnchor.constraint(equalTo: ytVideoPlayerView.centerXAnchor, constant: 90),
-//            showDanMuButton.centerYAnchor.constraint(equalTo: ytVideoPlayerView.centerYAnchor, constant: 150),
-//            showDanMuButton.widthAnchor.constraint(equalToConstant: 100),
-//            showDanMuButton.heightAnchor.constraint(equalToConstant: 40)
-//        ])
-//    }
-//    private func setPortraitAutoLayOut() {
-//        rootViewController.view.addSubview(tableView)
-//        removeAllContraints()
-//        NSLayoutConstraint.activate([
-//            baseView.leadingAnchor.constraint(equalTo: rootViewController.view.leadingAnchor),
-//            baseView.trailingAnchor.constraint(equalTo: rootViewController.view.trailingAnchor),
-//            baseView.topAnchor.constraint(equalTo: rootViewController.view.topAnchor),
-//            baseView.leftAnchor.constraint(equalTo: rootViewController.view.leftAnchor),
-//            baseView.widthAnchor.constraint(equalTo: rootViewController.view.widthAnchor),
-//            baseView.heightAnchor.constraint(equalTo: rootViewController.view.heightAnchor),
-//            ytVideoPlayerView.topAnchor.constraint(equalTo: rootViewController.view.topAnchor),
-//            ytVideoPlayerView.leftAnchor.constraint(equalTo: rootViewController.view.leftAnchor),
-//            ytVideoPlayerView.widthAnchor.constraint(equalTo: rootViewController.view.widthAnchor),
-//            ytVideoPlayerView.heightAnchor.constraint(equalTo: rootViewController.view.widthAnchor, multiplier: 9/16),
-//            baseView.topAnchor.constraint(equalTo: rootViewController.view.topAnchor),
-//            baseView.leftAnchor.constraint(equalTo: rootViewController.view.leftAnchor),
-//            baseView.widthAnchor.constraint(equalTo: rootViewController.view.widthAnchor),
-//            baseView.heightAnchor.constraint(equalTo: rootViewController.view.widthAnchor, multiplier: 9/16),
-//            tableView.topAnchor.constraint(equalTo: ytVideoPlayerView.bottomAnchor),
-//            tableView.leadingAnchor.constraint(equalTo: rootViewController.view.leadingAnchor),
-//            tableView.trailingAnchor.constraint(equalTo: rootViewController.view.trailingAnchor),
-//            tableView.bottomAnchor.constraint(equalTo: rootViewController.view.bottomAnchor),
-//            buttonsView.leadingAnchor.constraint(equalTo: rootViewController.view.leadingAnchor),
-//            buttonsView.trailingAnchor.constraint(equalTo: rootViewController.view.trailingAnchor),
-//            buttonsView.heightAnchor.constraint(equalTo: rootViewController.view.widthAnchor, multiplier: 9 / 16),
-//            buttonsView.topAnchor.constraint(equalTo: rootViewController.view.topAnchor),
-//            changeOrientationButton.centerXAnchor.constraint(equalTo: rootViewController.view.centerXAnchor),
-//            changeOrientationButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -16),
-//            changeOrientationButton.widthAnchor.constraint(equalToConstant: 30),
-//            changeOrientationButton.heightAnchor.constraint(equalToConstant: 30),
-//            danmuView.leadingAnchor.constraint(equalTo: ytVideoPlayerView.leadingAnchor),
-//            danmuView.trailingAnchor.constraint(equalTo: ytVideoPlayerView.trailingAnchor),
-//            danmuView.topAnchor.constraint(equalTo: ytVideoPlayerView.topAnchor),
-//            danmuView.heightAnchor.constraint(equalTo: ytVideoPlayerView.heightAnchor, multiplier: 5 / 10),
-//
-//            showDanMuTextFieldButton.trailingAnchor.constraint(equalTo: buttonsView.trailingAnchor, constant: -225),
-//            showDanMuTextFieldButton.bottomAnchor.constraint(equalTo: buttonsView.bottomAnchor, constant: -16),
-//            showDanMuTextFieldButton.widthAnchor.constraint(equalToConstant: 30),
-//            showDanMuTextFieldButton.heightAnchor.constraint(equalToConstant: 30),
-//            pauseButton.centerXAnchor.constraint(equalTo: buttonsView.centerXAnchor),
-//            pauseButton.centerYAnchor.constraint(equalTo: buttonsView.centerYAnchor),
-//            pauseButton.widthAnchor.constraint(equalToConstant: 50),
-//            pauseButton.heightAnchor.constraint(equalToConstant: 50),
-//
-//        ])
-//
+// MARK: -
+extension PlayerViewController {
+    
+    private func setupTableView() {
+        configureDataSource(tableView: tableView)
+        snapshot.appendSections([.main])
+        snapshot.appendItems([MKShow(image: "first", title: "first", playlistId: "first")])
+        snapshot.appendItems([MKShow(image: "second", title: "second", playlistId: "second")])
+        tableView.dataSource = dataSource
+        dataSource.apply(snapshot)
+    }
+    
+    private func configureDataSource(tableView: UITableView) {
+        dataSource = UITableViewDiffableDataSource<OneSection, MKShow>(
+            tableView: tableView,
+            cellProvider: { tableView, indexPath, item in
+                if indexPath.row == 1 {
+                    let cell = tableView.dequeueReusableCell(
+                        withIdentifier: ChatroomButtonTableViewCell.identifier,
+                        for: indexPath) as? ChatroomButtonTableViewCell
+                    guard let cell = cell else { return UITableViewCell() }
+                    cell.chatRoomButton.addTarget(self, action: #selector(self.showChatroom(sender:)), for: .touchUpInside)
+                    return cell
+                } else {
+                    return UITableViewCell()
+                }
+            }
+        )
+    }
+    // MARK: - Show Chatroom Button Action
+    @objc func showChatroom(sender: UIButton) {
+        let chatroomVC = ChatroomViewController()
+        if let sheet = chatroomVC.sheetPresentationController {
+            sheet.prefersGrabberVisible = true
+            sheet.detents = [.custom { _ in
+                600.0
+            }, .large()]
+            sheet.largestUndimmedDetentIdentifier = .large
+            chatroomVC.videoId = self.videoId
+            self.present(chatroomVC, animated: true)
+        }
+    }
+}
